PackageMetadata handler moved into a submodule
Move package metadata logic out of execution.rs into a new programmable_transactions/package_metadata.rs file.
