+++ iota-execution/latest/iota-adapter/src/programmable_transactions/execution.rs
@@ -21,7 +21,7 @@ mod checked {
         TypeTag, move_package::MovePackage,
     };
     use iota_types::{
-        IOTA_FRAMEWORK_PACKAGE_ID, auth_context,
+        auth_context,
         base_types::{
             IotaAddress, MoveLegacyTxContext, RESOLVED_ASCII_STR, RESOLVED_STD_OPTION,
             RESOLVED_UTF8_STR, TxContext, TxContextKind,
@@ -29,13 +29,13 @@ mod checked {
         coin::Coin,
         error::{ExecutionError, ExecutionErrorKind, IotaError, command_argument_error},
         execution_config_utils::to_binary_config,
-        id::{ID, RESOLVED_IOTA_ID},
+        id::RESOLVED_IOTA_ID,
         iota_sdk_types_conversions::type_tag_core_to_sdk,
         metrics::LimitsMetrics,
         move_package::{
-            IotaAttribute, MovePackageExt, PackageMetadata, RuntimeModuleMetadata,
-            RuntimeModuleMetadataWrapper, UpgradeCap, UpgradePolicy, UpgradeReceipt, UpgradeTicket,
-            normalize_deserialized_modules_with_metadata, normalize_modules_with_metadata,
+            IotaAttribute, MovePackageExt, RuntimeModuleMetadata, UpgradeCap, UpgradePolicy,
+            UpgradeReceipt, UpgradeTicket, normalize_deserialized_modules_with_metadata,
+            normalize_modules_with_metadata,
         },
         object::OBJECT_START_VERSION,
         storage::{PackageObject, get_package_objects},
@@ -53,15 +53,13 @@ mod checked {
         CompiledModule,
         compatibility::{Compatibility, InclusionCheck},
         errors::{Location, PartialVMResult, VMResult},
-        file_format::{
-            AbilitySet, CodeOffset, FunctionDefinitionIndex, LocalIndex, SignatureToken, Visibility,
-        },
-        file_format_common::{IOTA_METADATA_KEY, VERSION_6},
+        file_format::{AbilitySet, CodeOffset, FunctionDefinitionIndex, LocalIndex, Visibility},
+        file_format_common::VERSION_6,
         normalized,
     };
     use move_core_types::{
-        account_address::AccountAddress, ident_str, identifier::IdentStr,
-        language_storage::ModuleId, u256::U256,
+        account_address::AccountAddress, identifier::IdentStr, language_storage::ModuleId,
+        u256::U256,
     };
     use move_trace_format::format::MoveTraceBuilder;
     use move_vm_runtime::{
@@ -69,7 +67,7 @@ mod checked {
         session::{LoadedFunctionInstantiation, SerializedReturnValues},
     };
     use move_vm_types::loaded_data::runtime_types::{CachedDatatype, Type};
-    use serde::{Deserialize, Serialize, de::DeserializeSeed};
+    use serde::{Deserialize, de::DeserializeSeed};
     use tracing::instrument;
 
     use crate::{
@@ -80,12 +78,8 @@ mod checked {
             ensure_serialized_size,
         },
         gas_charger::GasCharger,
-        programmable_transactions::context::*,
+        programmable_transactions::{context::*, package_metadata::*},
     };
-
-    const CREATE_PACKAGE_METADATA_V1_WITH_DYNAMIC_METADATA_FN_NAME: &IdentStr =
-        ident_str!("create_package_metadata_v1_with_dynamic_metadata");
-    const PACKAGE_METADATA_MODULE_NAME: &IdentStr = ident_str!("package_metadata");
 
     /// Executes a `ProgrammableTransaction` in the specified `ExecutionMode`,
     /// applying a series of commands to the execution context. The
@@ -966,356 +960,6 @@ mod checked {
                 ))
             }
             Ok(Ok(pkgs)) => Ok(pkgs),
-        }
-    }
-
-    #[derive(Copy, Clone)]
-    enum PackageMetadataHandler {
-        V1,
-        Dynamic,
-    }
-
-    impl PackageMetadataHandler {
-        fn from_protocol_config(protocol_config: &ProtocolConfig) -> Self {
-            if protocol_config.package_metadata_with_dynamic_module_metadata() {
-                Self::Dynamic
-            } else {
-                Self::V1
-            }
-        }
-
-        fn supports_view_function_metadata(self) -> bool {
-            matches!(self, Self::Dynamic)
-        }
-
-        fn should_publish_package(
-            self,
-            modules_metadata: &BTreeMap<String, PendingModuleMetadata>,
-        ) -> bool {
-            match self {
-                Self::V1 => modules_metadata
-                    .values()
-                    .any(|md| !md.authenticator_metadata.is_empty()),
-                Self::Dynamic => modules_metadata.values().any(|md| !md.is_empty()),
-            }
-        }
-
-        fn publish_package(
-            self,
-            context: &mut ExecutionContext<'_, '_, '_>,
-            modules_metadata: &BTreeMap<String, PendingModuleMetadata>,
-            storage_id: ObjectId,
-            runtime_id: ObjectId,
-            package_version: u64,
-            trace_builder_opt: &mut Option<MoveTraceBuilder>,
-        ) -> Result<(), ExecutionError> {
-            match self {
-                Self::V1 => self.publish_v1(
-                    context,
-                    modules_metadata,
-                    storage_id,
-                    runtime_id,
-                    package_version,
-                ),
-                Self::Dynamic => self.publish_v1_with_dynamic_metadata(
-                    context,
-                    modules_metadata,
-                    storage_id,
-                    runtime_id,
-                    package_version,
-                    trace_builder_opt,
-                ),
-            }
-        }
-
-        fn publish_v1(
-            self,
-            context: &mut ExecutionContext<'_, '_, '_>,
-            modules_metadata: &BTreeMap<String, PendingModuleMetadata>,
-            storage_id: ObjectId,
-            runtime_id: ObjectId,
-            package_version: u64,
-        ) -> Result<(), ExecutionError> {
-            let metadata_uid = context.package_derived_metadata_id(storage_id)?;
-            // Create the package metadata object content
-            let modules_metadata_v1 = modules_metadata
-                .iter()
-                .map(|(mod_name, pending_metadata)| {
-                    (
-                        mod_name.clone(),
-                        pending_metadata.authenticator_metadata.clone(),
-                    )
-                })
-                .collect();
-            let metadata = PackageMetadata::new_v1(
-                metadata_uid,
-                storage_id,
-                runtime_id,
-                package_version,
-                modules_metadata_v1,
-            );
-            // Turn the content into an object
-            let package_metadata = context.make_object_value(
-                metadata.type_(),
-                // used_in_non_entry_move_call
-                false,
-                &metadata.to_bcs_bytes(),
-            )?;
-            // Freeze the package metadata object
-            context.freeze_object(package_metadata)
-        }
-
-        fn publish_v1_with_dynamic_metadata(
-            self,
-            context: &mut ExecutionContext<'_, '_, '_>,
-            modules_metadata: &BTreeMap<String, PendingModuleMetadata>,
-            storage_id: ObjectId,
-            runtime_id: ObjectId,
-            package_version: u64,
-            trace_builder_opt: &mut Option<MoveTraceBuilder>,
-        ) -> Result<(), ExecutionError> {
-            let (modules, auth_functions, type_names, view_function_names) =
-                package_metadata_constructor_args(modules_metadata);
-            // The Move constructor derives both the package metadata object address
-            // and the per-module metadata object addresses from the package storage
-            // ID, so we pass the storage ID as the package ID.
-            let args = vec![
-                to_bcs_argument(&ID::new(storage_id), "package metadata package ID")?,
-                to_bcs_argument(&ID::new(storage_id), "package metadata storage ID")?,
-                to_bcs_argument(&ID::new(runtime_id), "package metadata runtime ID")?,
-                to_bcs_argument(&package_version, "package metadata version")?,
-                to_bcs_argument(&modules, "package metadata modules")?,
-                to_bcs_argument(&auth_functions, "package metadata authenticator functions")?,
-                to_bcs_argument(&type_names, "package metadata type names")?,
-                to_bcs_argument(&view_function_names, "package metadata view functions")?,
-            ];
-            execute_package_metadata_constructor(
-                context,
-                CREATE_PACKAGE_METADATA_V1_WITH_DYNAMIC_METADATA_FN_NAME,
-                args,
-                trace_builder_opt,
-            )
-        }
-    }
-
-    #[derive(Default)]
-    struct PendingModuleMetadata {
-        authenticator_metadata: BTreeMap<String, TypeTag>,
-        view_function_metadata: Vec<String>,
-    }
-
-    impl PendingModuleMetadata {
-        fn is_empty(&self) -> bool {
-            self.authenticator_metadata.is_empty() && self.view_function_metadata.is_empty()
-        }
-    }
-
-    /// Creates package metadata for a Move package by extracting module
-    /// metadata and passing it to the framework package metadata constructor.
-    /// The framework constructor builds and freezes the metadata object. If no
-    /// relevant metadata is found, the function exits without creating any
-    /// package metadata.
-    fn create_and_freeze_package_metadata_if_present(
-        context: &mut ExecutionContext<'_, '_, '_>,
-        modules: &[CompiledModule],
-        storage_id: ObjectId,
-        runtime_id: ObjectId,
-        package_version: u64,
-        trace_builder_opt: &mut Option<MoveTraceBuilder>,
-    ) -> Result<(), ExecutionError> {
-        let package_metadata_handler =
-            PackageMetadataHandler::from_protocol_config(context.protocol_config);
-        let mut modules_metadata_map = BTreeMap::new();
-        // Extract metadata for each module
-        for module in modules {
-            if let Some(md) = module
-                .metadata
-                .iter()
-                .find(|md| md.key == IOTA_METADATA_KEY.to_vec())
-            {
-                // At this point, if the metadata is present, it should have been already
-                // validated by the iota-verifier during package verification (in
-                // `publish_and_verify_modules`).
-                let runtime_module_metadata: RuntimeModuleMetadata =
-                    bcs::from_bytes::<RuntimeModuleMetadataWrapper>(&md.value)
-                        .map_err(|_| {
-                            ExecutionError::from_kind(
-                                ExecutionErrorKind::VmVerificationOrDeserializationError,
-                            )
-                        })?
-                        .try_into()
-                        .map_err(|_| {
-                            ExecutionError::from_kind(
-                                ExecutionErrorKind::VmVerificationOrDeserializationError,
-                            )
-                        })?;
-
-                // Process functions for each module in order to create package
-                // metadata.
-                let mut pending_module_metadata = PendingModuleMetadata::default();
-                for (fn_name, fn_attributes) in runtime_module_metadata.fun_attributes_iter() {
-                    // Check attributes
-                    for attribute in fn_attributes {
-                        match attribute {
-                            IotaAttribute::Authenticator(attribute) if attribute.version == 1 => {
-                                let contains =
-                                    pending_module_metadata.authenticator_metadata.insert(
-                                        fn_name.to_string(),
-                                        get_authenticator_first_param_type_tag(module, &fn_name)?,
-                                    );
-                                debug_assert!(
-                                    contains.is_none(),
-                                    "Duplicate function metadata for authenticator"
-                                );
-                            }
-                            IotaAttribute::View
-                                if package_metadata_handler.supports_view_function_metadata() =>
-                            {
-                                pending_module_metadata
-                                    .view_function_metadata
-                                    .push(fn_name.to_string());
-                            }
-                            _ => { /* Other attributes are ignored. */ }
-                        }
-                    }
-                }
-                // Fill the package metadata with a module handle (and its related function
-                // metadata) only if there is at least one function with
-                // relevant metadata
-                if !pending_module_metadata.is_empty() {
-                    modules_metadata_map.insert(module.name().to_string(), pending_module_metadata);
-                }
-            }
-        }
-
-        // Only publish package metadata if there is at least one module with
-        // relevant metadata
-        if package_metadata_handler.should_publish_package(&modules_metadata_map) {
-            package_metadata_handler.publish_package(
-                context,
-                &modules_metadata_map,
-                storage_id,
-                runtime_id,
-                package_version,
-                trace_builder_opt,
-            )?;
-        }
-        Ok(())
-    }
-
-    fn package_metadata_constructor_args(
-        modules_metadata: &BTreeMap<String, PendingModuleMetadata>,
-    ) -> (
-        Vec<String>,
-        Vec<Vec<String>>,
-        Vec<Vec<String>>,
-        Vec<Vec<String>>,
-    ) {
-        let mut modules = Vec::with_capacity(modules_metadata.len());
-        let mut auth_functions = Vec::with_capacity(modules_metadata.len());
-        let mut type_names = Vec::with_capacity(modules_metadata.len());
-        let mut view_function_names = Vec::with_capacity(modules_metadata.len());
-
-        for (module_name, metadata) in modules_metadata {
-            modules.push(module_name.clone());
-
-            let mut module_auth_functions =
-                Vec::with_capacity(metadata.authenticator_metadata.len());
-            let mut module_type_names = Vec::with_capacity(metadata.authenticator_metadata.len());
-            for (function_name, account_type) in &metadata.authenticator_metadata {
-                module_auth_functions.push(function_name.clone());
-                module_type_names.push(account_type.to_canonical_string(false));
-            }
-            auth_functions.push(module_auth_functions);
-            type_names.push(module_type_names);
-            view_function_names.push(metadata.view_function_metadata.clone());
-        }
-
-        (modules, auth_functions, type_names, view_function_names)
-    }
-
-    fn execute_package_metadata_constructor(
-        context: &mut ExecutionContext<'_, '_, '_>,
-        function: &IdentStr,
-        args: Vec<Vec<u8>>,
-        trace_builder_opt: &mut Option<MoveTraceBuilder>,
-    ) -> Result<(), ExecutionError> {
-        let saved_linkage = context.linkage_view.steal_linkage();
-        let result = (|| {
-            let original_address = context.set_link_context(IOTA_FRAMEWORK_PACKAGE_ID)?;
-            let runtime_id =
-                ModuleId::new(original_address, PACKAGE_METADATA_MODULE_NAME.to_owned());
-            context
-                .execute_function_bypass_visibility(
-                    &runtime_id,
-                    function,
-                    vec![],
-                    args,
-                    trace_builder_opt,
-                )
-                .map_err(|e| context.convert_vm_error(e))
-        })();
-        context.linkage_view.reset_linkage();
-        let res = result.and_then(|result| {
-            let SerializedReturnValues {
-                mutable_reference_outputs,
-                return_values,
-            } = result;
-            assert_invariant!(
-                return_values.is_empty(),
-                "package metadata constructor should not have return values"
-            );
-            assert_invariant!(
-                mutable_reference_outputs.is_empty(),
-                "package metadata constructor should not have mutable reference outputs"
-            );
-            Ok(())
-        });
-        context.linkage_view.restore_linkage(saved_linkage)?;
-        res
-    }
-
-    fn to_bcs_argument<T: Serialize>(
-        value: &T,
-        description: &'static str,
-    ) -> Result<Vec<u8>, ExecutionError> {
-        bcs::to_bytes(value).map_err(|e| {
-            ExecutionError::invariant_violation(format!("failed to serialize {description}: {e}"))
-        })
-    }
-
-    fn get_authenticator_first_param_type_tag(
-        module: &CompiledModule,
-        authenticate_fn_name: &impl AsRef<str>,
-    ) -> Result<TypeTag, ExecutionError> {
-        // Entering into this function, the verifier must have already been run,
-        // so we can assume the function exists and has the correct signature.
-        let Some((_, fn_definition)) = module.find_function_def_by_name(authenticate_fn_name)
-        else {
-            return Err(ExecutionError::from_kind(
-                ExecutionErrorKind::VmInvariantViolation,
-            ));
-        };
-        let fn_handle = module.function_handle_at(fn_definition.function);
-        let fn_signature = module.signature_at(fn_handle.parameters);
-        // We need the first parameter to be a reference type so we can extract the
-        // inner as the type tag.
-        match &fn_signature.0[0] {
-            SignatureToken::Reference(ref_param) => {
-                let pool = &mut normalized::RcPool::new();
-                if let Some(type_tag) =
-                    normalized::Type::new(pool, module, ref_param).to_type_tag(pool)
-                {
-                    Ok(type_tag_core_to_sdk(&type_tag))
-                } else {
-                    Err(ExecutionError::from_kind(
-                        ExecutionErrorKind::VmVerificationOrDeserializationError,
-                    ))
-                }
-            }
-            _ => Err(ExecutionError::from_kind(
-                ExecutionErrorKind::VmVerificationOrDeserializationError,
-            )),
         }
     }
