+++ iota-execution/latest/iota-adapter/src/programmable_transactions/mod.rs
@@ -5,3 +5,4 @@
 pub mod context;
 pub mod execution;
 pub mod linkage_view;
+pub mod package_metadata;
